requirements.txt
pandas 
numpy
matplotlib 
python-can 
canlib